Removed EmailerAgent from examples and updated in ops/EmailerAgent
