--- examples/EmailerAgent/setup.py
@@ -1,75 +1,0 @@
-# -*- coding: utf-8 -*- {{{
-# vim: set fenc=utf-8 ft=python sw=4 ts=4 sts=4 et:
-#
-# Copyright 2017, Battelle Memorial Institute.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# This material was prepared as an account of work sponsored by an agency of
-# the United States Government. Neither the United States Government nor the
-# United States Department of Energy, nor Battelle, nor any of their
-# employees, nor any jurisdiction or organization that has cooperated in the
-# development of these materials, makes any warranty, express or
-# implied, or assumes any legal liability or responsibility for the accuracy,
-# completeness, or usefulness or any information, apparatus, product,
-# software, or process disclosed, or represents that its use would not infringe
-# privately owned rights. Reference herein to any specific commercial product,
-# process, or service by trade name, trademark, manufacturer, or otherwise
-# does not necessarily constitute or imply its endorsement, recommendation, or
-# favoring by the United States Government or any agency thereof, or
-# Battelle Memorial Institute. The views and opinions of authors expressed
-# herein do not necessarily state or reflect those of the
-# United States Government or any agency thereof.
-#
-# PACIFIC NORTHWEST NATIONAL LABORATORY operated by
-# BATTELLE for the UNITED STATES DEPARTMENT OF ENERGY
-# under Contract DE-AC05-76RL01830
-# }}}
-
-from setuptools import setup, find_packages
-from os import path
-
-MAIN_MODULE = 'agent'
-
-# Find the agent package that contains the main module
-packages = find_packages('.')
-agent_package = ''
-for package in find_packages():
-    # Because there could be other packages such as tests
-    if path.isfile(package + '/' + MAIN_MODULE + '.py') is True:
-        agent_package = package
-if not agent_package:
-    raise RuntimeError('None of the packages under {dir} contain the file '
-                       '{main_module}'.format(main_module=MAIN_MODULE + '.py',
-                                              dir=path.abspath('.')))
-
-# Find the version number from the main module
-agent_module = agent_package + '.' + MAIN_MODULE
-_temp = __import__(agent_module, globals(), locals(), ['__version__'], -1)
-__version__ = _temp.__version__
-
-# Setup
-setup(
-    name=agent_package + 'agent',
-    version=__version__,
-    author_email="rajveer.singh@pnnl.gov",
-    description="Sends email on failover",
-    author="Raj",
-    install_requires=['volttron'],
-    packages=packages,
-    entry_points={
-        'setuptools.installation': [
-            'eggsecutable = ' + agent_module + ':main',
-        ]
-    }
-)
